Change default sigma from 1.5 to 1.6.
This seemingly meaningless change drastically improves the number of
correct matches when an image is scaled roughly by the factor of 1.4.
This is because the width of descriptor window now expands
progressively from 2x16 to 4x16 on different scales rather than
being 3x16 most of the time.

TODO: I can play with different window widths to improve matching
even more.